Fix an error on linux-grsec
Using perl regexes in grep requires RWX anonymous mmap which is
prohibited by the linux-grsec kernel.  Normally, special flags are
added to binaries that need these mappings but that is not the case
for grep (people don't normally use perl regex's in grep).  By using
perl instead, users don't have to add the flag themselves.  This is
not a huge issue since perl is almost always installed on a user's
system.