Allow prepare() to modify pkgver()
See discussion in #645. Originally, (in 40e4f07) --noprepare was removed
from makepkg during the integrity check code because prepare() could
change the pkgver. This caused another bug, that prepare() was performed
multiple times, causing things like patching to fail, so this was
reverted in 71d698e.

The proper fix is to *only* run prepare() during the early phase, and
use the results in the build.

This was tested with `linux-cdown`, which builds successfully with this
patch.