Proceed: use `read -s`, wrapped in while loop
This makes `pacaur` handle unexpected input better, e.g. when a cursor
key is used.

TODO: It is not clear to me why `$cleancache` is handled in a special
way?! (like the dumb terminal, added in 5d5c502)
Is it meant to reflect pacman's prompt?  Is this still the case there?